+++ infrastructure/project/variables.tf
@@ -29,8 +29,8 @@ variable "sensitive_custom_env_vars" {
 variable "env_vars" {
   type = object({
     shared = optional(object({
-      NEXT_PUBLIC_CLERK_SIGN_IN_URL   = optional(string)
-      NEXT_PUBLIC_CLERK_SIGN_UP_URL   = optional(string)
+      NEXT_PUBLIC_CLERK_SIGN_IN_URL   = optional(string, "/sign-in")
+      NEXT_PUBLIC_CLERK_SIGN_UP_URL   = optional(string, "/sign-in")
       NEXT_PUBLIC_CLIENT_APP_BASE_URL = optional(string)
     }))
     prod = optional(object({